refactor: useProduct 훅 분리
